training/requirements.txt
unsloth
trl
peft
datasets
accelerate
bitsandbytes
transformers
torch
python-dotenv
openai
